Moved ErrorResponse into model module
